add gis

+++ src/component/meter-form/action.ts
@@ -2,6 +2,8 @@
 import { prisma } from "@/prisma/prisma-client";
 import { InstallationFormData } from "@/shema/instaltion-form";
 import { MeterFormData } from "@/shema/meter-form";
+import { EsriPointGeometry, PeaMeterDetailAttributes } from "@/types/gis";
+import { PhaseInstallation, PhaseSystem } from "@prisma/client";
 import sharp from "sharp";
 import { pick } from "zod/v4-mini";
 
@@ -12,6 +14,8 @@ export const getMeter = async (id: string) => {
     },
   });
 
+  console.log(meterData);
+
   if (!meterData) throw new Error("Meter not found");
   return meterData;
 };
@@ -21,6 +25,7 @@ export async function submitMeterForm(formData: MeterFormData) {
 
   const result = await prisma.meter.create({
     data: {
+      ca: formData.ca,
       peaNoNew: formData.peaNoNew,
       pickerName: formData.issuerName,
       pickerDate: formData.issueDate,
@@ -50,13 +55,13 @@ export async function updateMeterForm(
     where: {
       id,
     },
-    data : {
+    data: {
       peaNoNew: formData.peaNoNew,
       pickerName: formData.issuerName,
       pickerDate: formData.issueDate,
       newMeterImageId: data.newMeterImageId,
       newMeterImageUrl: data.newMeterImageUrl,
-    }
+    },
   });
 
   return result;
@@ -70,16 +75,18 @@ export async function submitInstallationForm(
   }: { isRemoveNewImage: boolean; isRemoveOldImage: boolean }
 ) {
   let data = await getMeter(formData.id);
-  if (isRemoveOldImage) {
-    if(data.oldMeterImageId){
-      await deleteImage(data.oldMeterImageId);
-    }
+  if (isRemoveOldImage && data.oldMeterImageId) {
+    await deleteImage(data.oldMeterImageId);
+  }
+
+  if (isRemoveOldImage || !data.oldMeterImageId) {
     const oldMeterUpload = await uploadImage(formData.oldMeterImage!);
     data.oldMeterImageId = oldMeterUpload.fileId;
     data.oldMeterImageUrl = oldMeterUpload.url;
   }
+
   if (isRemoveNewImage) {
-    if(data.newMeterImageId){
+    if (data.newMeterImageId) {
       await deleteImage(data.newMeterImageId);
     }
     const newMeterUpload = await uploadImage(formData.newMeterImage!);
@@ -92,6 +99,7 @@ export async function submitInstallationForm(
       id: formData.id,
     },
     data: {
+      ca: formData.ca,
       peaNoNew: formData.peaNoNew,
       peaNoOld: formData.peaNoOld,
       installationName: formData.installationName,
@@ -173,4 +181,108 @@ export async function deleteImage(fileId: string) {
 
 export async function mockSubmit(data: InstallationFormData) {
   console.log(data);
+}
+
+export async function updateFromGis(
+  data: PeaMeterDetailAttributes,
+  point: EsriPointGeometry,
+  id: string
+) {
+  const customerName = `${data["PEA.METER_DETAIL.PREFIX"] || ""}${
+    data["PEA.METER_DETAIL.CUSTOMERNAME"] || ""
+  } ${data["PEA.METER_DETAIL.CUSTOMERSIRNAME"] || ""}`;
+  let address = "";
+  if (data["PEA.METER_DETAIL.VILLAGEBUILDING"]) {
+    address += data["PEA.METER_DETAIL.VILLAGEBUILDING"];
+  }
+
+  if (data["PEA.METER_DETAIL.FLOORNO"])
+    address += ` ${data["PEA.METER_DETAIL.FLOORNO"]}`;
+  if (data["PEA.METER_DETAIL.ROOMNO"])
+    address += ` ${data["PEA.METER_DETAIL.ROOMNO"]}`;
+  if (data["PEA.METER_DETAIL.ADDRESSNO"])
+    address += ` ${data["PEA.METER_DETAIL.ADDRESSNO"]}`;
+  if (data["PEA.METER_DETAIL.MOO"])
+    address += ` ${data["PEA.METER_DETAIL.MOO"]}`;
+  if (data["PEA.METER_DETAIL.TROK"])
+    address += ` ${data["PEA.METER_DETAIL.TROK"]}`;
+  if (data["PEA.METER_DETAIL.SOI"])
+    address += ` ${data["PEA.METER_DETAIL.SOI"]}`;
+  if (data["PEA.METER_DETAIL.STREET"])
+    address += ` ${data["PEA.METER_DETAIL.STREET"]}`;
+  if (data["PEA.METER_DETAIL.TUMBOL"])
+    address += ` ${data["PEA.METER_DETAIL.TUMBOL"]}`;
+  if (data["PEA.METER_DETAIL.AMPHOE"])
+    address += ` ${data["PEA.METER_DETAIL.AMPHOE"]}`;
+  if (data["PEA.METER_DETAIL.CHANGWAT"])
+    address += ` ${data["PEA.METER_DETAIL.CHANGWAT"]}`;
+  if (data["PEA.METER_DETAIL.POSTCODE"])
+    address += ` ${data["PEA.METER_DETAIL.POSTCODE"]}`;
+
+  const materialNumberOld = data["PEA.METER_DETAIL.MATERIALNUMBER"];
+  const metertypeOld = data["PEA.METER_DETAIL.METERTYPE"];
+  const peaCode = data["PEA.METER_DETAIL.CODE"];
+
+  let phaseSystem: string | undefined = undefined;
+  if (data["PEA.DS_LowVoltageMeter.SUBTYPECODE"])
+    switch (data["PEA.DS_LowVoltageMeter.SUBTYPECODE"]) {
+      case 1:
+        phaseSystem = "ONEPHASE";
+        break;
+      case 3:
+        phaseSystem = "THREEPHASE";
+        break;
+      default:
+        break;
+    }
+
+  let phaseInstallation: string | undefined = undefined;
+  if (data["PEA.DS_LowVoltageMeter.PHASEDESIGNATION"]) {
+    switch (data["PEA.DS_LowVoltageMeter.PHASEDESIGNATION"]) {
+      case 1:
+        phaseInstallation = "CN";
+        break;
+      case 2:
+        phaseInstallation = "BN";
+        break;
+      case 4:
+        phaseInstallation = "AN";
+        break;
+      case 7:
+        phaseInstallation = "ABCN";
+        break;
+      default:
+        break;
+    }
+  }
+
+  if (data["PEA.DS_GroupMeter_Detail.PHASEDESIGNATION"])
+    phaseInstallation = data["PEA.DS_GroupMeter_Detail.PHASEDESIGNATION"];
+
+  const result = await prisma.meter.update({
+    where: {
+      id,
+    },
+    data: {
+      customerName,
+      customerAddress: address,
+      materialNumberOld,
+      metertypeOld,
+      peaCode,
+      phaseSystem,
+      phaseInstallation,
+      meterOldLocation: `${point.y}, ${point.x}`
+    },
+  });
+
+  return result;
+}
+
+export async function exportIsInstalledAndNoDataFromGIS() {
+  const data = await prisma.meter.findMany({
+    where: {
+      peaNoOld: { isSet: true },
+    },
+  });
+  return data
 }
